CBP - Stocks Transactions - Cash Money Destock - QA Fix GetManager #3026
