+++ CBP/plugin/stock_transactions/fermat-cbp-plugin-stock-transactions-crypto-money-destock-bitdubai/src/main/java/com/bitdubai/fermat_cbp_plugin/layer/stock_transactions/crypto_money_destock/developer/bitdubai/version_1/structure/StockTransactionCryptoMoneyDestockManager.java
@@ -1,14 +1,20 @@
 package com.bitdubai.fermat_cbp_plugin.layer.stock_transactions.crypto_money_destock.developer.bitdubai.version_1.structure;
 
+import com.bitdubai.fermat_api.layer.all_definition.enums.CryptoCurrency;
 import com.bitdubai.fermat_api.layer.all_definition.exceptions.InvalidParameterException;
 import com.bitdubai.fermat_api.layer.osa_android.database_system.DatabaseTableFilter;
 import com.bitdubai.fermat_api.layer.osa_android.database_system.PluginDatabaseSystem;
 import com.bitdubai.fermat_cbp_api.all_definition.business_transaction.BankMoneyTransaction;
 import com.bitdubai.fermat_cbp_api.all_definition.business_transaction.CryptoMoneyTransaction;
+import com.bitdubai.fermat_cbp_api.all_definition.enums.OriginTransaction;
+import com.bitdubai.fermat_cbp_api.all_definition.enums.TransactionStatusRestockDestock;
+import com.bitdubai.fermat_cbp_api.layer.stock_transactions.crypto_money_destock.exceptions.CantCreateCryptoMoneyDestockException;
+import com.bitdubai.fermat_cbp_api.layer.stock_transactions.crypto_money_destock.interfaces.CryptoMoneyDestockManager;
 import com.bitdubai.fermat_cbp_plugin.layer.stock_transactions.crypto_money_destock.developer.bitdubai.version_1.database.StockTransactionsCryptoMoneyDestockDatabaseDao;
 import com.bitdubai.fermat_cbp_plugin.layer.stock_transactions.crypto_money_destock.developer.bitdubai.version_1.exceptions.DatabaseOperationException;
 import com.bitdubai.fermat_cbp_plugin.layer.stock_transactions.crypto_money_destock.developer.bitdubai.version_1.exceptions.MissingCryptoMoneyDestockDataException;
 
+import java.sql.Timestamp;
 import java.util.List;
 import java.util.UUID;
 
@@ -18,7 +24,8 @@ import java.util.UUID;
  *
  * Created by franklin on 17/11/15.
  */
-public class StockTransactionCryptoMoneyDestockManager {
+public class StockTransactionCryptoMoneyDestockManager implements
+        CryptoMoneyDestockManager {
     private final PluginDatabaseSystem pluginDatabaseSystem;
     private final UUID pluginId;
 
@@ -34,16 +41,43 @@ public class StockTransactionCryptoMoneyDestockManager {
         this.pluginId             = pluginId            ;
     }
 
+    @Override
+    public void createTransactionDestock(String publicKeyActor, CryptoCurrency cryptoCurrency, String cbpWalletPublicKey, String cryWalletPublicKey, float amount, String memo, float priceReference, OriginTransaction originTransaction) throws CantCreateCryptoMoneyDestockException {
+        java.util.Date date = new java.util.Date();
+        Timestamp timestamp = new Timestamp(date.getTime());
+        CryptoMoneyDestockTransactionImpl cryptoMoneyRestockTransaction = new CryptoMoneyDestockTransactionImpl(
+                UUID.randomUUID(),
+                publicKeyActor,
+                cryptoCurrency,
+                cbpWalletPublicKey,
+                cryWalletPublicKey,
+                memo,
+                "INIT TRANSACTION",
+                amount,
+                timestamp,
+                TransactionStatusRestockDestock.INIT_TRANSACTION,
+                priceReference,
+                originTransaction);
+
+        try {
+            saveCryptoMoneyDestockTransactionData(cryptoMoneyRestockTransaction);
+        } catch (DatabaseOperationException e) {
+            e.printStackTrace();
+        } catch (MissingCryptoMoneyDestockDataException e) {
+            e.printStackTrace();
+        }
+    }
     private StockTransactionsCryptoMoneyDestockDatabaseDao getStockTransactionCryptoMoneyDestockDao() {
 
         return new StockTransactionsCryptoMoneyDestockDatabaseDao(pluginDatabaseSystem, pluginId);
     }
 
-    public void saveCryptoMoneyDestockTransactionData(CryptoMoneyTransaction cryptoMoneyTransaction) throws DatabaseOperationException, MissingCryptoMoneyDestockDataException {
+    private void saveCryptoMoneyDestockTransactionData(CryptoMoneyTransaction cryptoMoneyTransaction) throws DatabaseOperationException, MissingCryptoMoneyDestockDataException {
         getStockTransactionCryptoMoneyDestockDao().saveCryptoMoneyDestockTransactionData(cryptoMoneyTransaction);
     }
 
     public List<CryptoMoneyTransaction> getCryptoMoneyTransactionList(DatabaseTableFilter filter) throws DatabaseOperationException, InvalidParameterException{
         return getStockTransactionCryptoMoneyDestockDao().getCryptoMoneyTransactionList(filter);
     }
+
 }
